new method of checking if already visited
takes the generated list with word counts and removes everything but the links. checks if the url, defragged, is then in visited_urls (all_webpage_count.txt). if it is, returns an empty list, else continues the scraping process.

done primarily because testing causes many restarts, which resets the visited_urls each time, thus creating an infinite crawl